Uploading Design File & Structing The Project Files
I uploaded in the folder "design" the design file on which the whole project is based.
In other words, we will recreate the design to an actual HTML form.

I added an empty css file, "styles.css", in the folder "css" where the CSS rules will be added.
I created an empty HTML file, "index.html", which will contain the structure of our HTML form.