Uploading Resources & Modifying HTML & CSS
The images that are about to get used on this project are inserted in the "img" folder,
as well as the fonts that will be used for the ODIN logo section.

The "styles.css" was modified to add the custom font using Font Squirrel, and
The "index.html" file was modified adding the HTML boilerplate and linking the CSS file as well.