Styling The First Column
The first column, which consists of the background image, the odin logo and the attribution of the
background image, is finished.

The div containing the odin logo and the text "ODIN" was positioned using "absolute" positioning.
Thus, it is no longer part of the normal document flow.

The attribution was placed at the end of the div using flexbox (flex-end).